conformance: better classify SERVFAIL test
