+++ conformance/packages/conformance-tests/src/resolver/dnssec/scenarios.rs
@@ -1,3 +1,2 @@
-mod bogus;
 mod ede;
 mod secure;
